More explicit linking
Newer versions of gcc apply the --as-needed linking flag automatically,
with this flag the linker only includes libraries when they are resolved
as a dependency of the executable. As this option was not previously
given the ordering of linking dependecies didn't matter, however it now
does. This refactors the Makefiles so that the ordering of dependencies is
given, and adds the --no-undefined flag so that compilation will not
succed unless all dependecies are specified.